resolved some minor issues

+++ src/app/api/testinomials/[id]/route.js
@@ -2,10 +2,14 @@ import { NextResponse } from "next/server";
 import { connectDB } from "@/lib/mongodb";
 import Testimonial from "@/models/Testinomials";
 
-export async function PATCH(req, { params }) {
+export async function PATCH(req, context) {
   try {
     await connectDB();
-    const { id } = params;
+    const user = await verifyAdmin();
+    if (!user || user.role !== "admin") {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    const { id } = await context.params;
     const body = await req.json();
     const updated = await Testimonial.findByIdAndUpdate(id, body, {
       new: true,
@@ -16,10 +20,14 @@ export async function PATCH(req, { params }) {
   }
 }
 
-export async function DELETE(req, { params }) {
+export async function DELETE(req, context) {
   try {
     await connectDB();
-    const { id } = params;
+    const user = await verifyAdmin();
+    if (!user || user.role !== "admin") {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    const { id } = await context.params;
     await Testimonial.findByIdAndDelete(id);
     return NextResponse.json({ message: "Deleted successfully" });
   } catch {
